new algo - matrix for cfqp
