+++ project/graph_utils.py
@@ -1,10 +1,8 @@
-from collections import defaultdict, deque
-from itertools import product
 from typing import Tuple, NamedTuple, Set, Union, Iterable, Any
 
 import networkx.drawing.nx_pydot as nx_pydot
-from networkx import MultiDiGraph, read_graphml
-from pyformlang.cfg import CFG, Terminal, Variable
+from networkx import MultiDiGraph
+from pyformlang.cfg import CFG
 import cfpq_data
 
 from project.cfg_utils import to_weak_cfg, cfg_from_text, cfg_from_file
@@ -105,128 +103,3 @@ def create_and_save_labeled_two_cycles_graph(
 def get_cnf(cfg: CFG) -> CFG:
     weak_cfg = to_weak_cfg(cfg)
     return weak_cfg.to_normal_form()
-
-
-def hellings_algorithm(
-    cfg: Union[str, CFG], graph: MultiDiGraph
-) -> Set[Tuple[str, str, str]]:
-    """
-    Computes the reachability information for all pairs of vertices in the given graph and context-free grammar.
-
-    Args:
-        cfg(CFG): The context-free grammar (CFG) to use for reachability analysis.
-        graph(MultiDiGraph): The directed graph on which to perform reachability analysis.
-
-    Returns:
-        Set[Tuple[int, Variable, int]]: A set of triples (start_vertex, nonterminal, end_vertex)
-        representing the reachability information for all pairs of vertices in the graph.
-    """
-    if isinstance(cfg, str):
-        cfg = cfg_from_text(cfg)
-    if graph.number_of_nodes() == 0:
-        return set()
-
-    wcnf = to_weak_cfg(cfg)
-    term_to_nonterm = defaultdict(set)
-    nonterm_pair_to_nonterm = defaultdict(set)
-    epsilon_nonterms = set()
-    for prod in wcnf.productions:
-        head, body = prod.head, prod.body
-        body_len = len(body)
-        if body_len == 0:
-            epsilon_nonterms.add(head)
-        elif body_len == 1:
-            term_to_nonterm[head].add(body[0])
-        elif body_len == 2:
-            nonterm_pair_to_nonterm[head].add((body[0], body[1]))
-
-    result = {(node, var, node) for node in graph.nodes for var in epsilon_nonterms}
-    result.update(
-        (i, n, j)
-        for i, j, label in graph.edges(data="label")
-        for n, terms in term_to_nonterm.items()
-        if Terminal(label) in terms
-    )
-
-    while True:
-        new_result = result.copy()
-        for i, j, k in product(graph.nodes, repeat=3):
-            for var1, pairs in nonterm_pair_to_nonterm.items():
-                for var2, var3 in pairs:
-                    if (i, var2, j) in result and (j, var3, k) in result:
-                        new = (i, var1, k)
-                        if new not in result:
-                            new_result.add(new)
-        if new_result == result:
-            break
-        result = new_result
-
-    return result
-
-
-def hellings_algorithm_from_file(
-    cfg: Union[str, CFG], path_to_graph: str
-) -> Set[Tuple[str, str, str]]:
-    """
-    Runs the Hellings algorithm on a graph read from a file.
-
-    The file should contain one edge per line, with each line formatted as "source_node target_node edge_label".
-
-    Args:
-        cfg: The context-free grammar to use for the algorithm, either as a string or as a CFG object.
-        path_to_graph: The path to the file containing the graph.
-
-    Returns:
-        The set of all valid paths through the graph that match the grammar, as tuples of (start_node, end_node, label).
-    """
-    graph = MultiDiGraph()
-    with open(path_to_graph, "r") as f:
-        for line in f:
-            line = line.strip()
-            if line:
-                source_node, target_node, edge_label = line.split()
-                graph.add_edge(int(source_node), int(target_node), label=edge_label)
-    return hellings_algorithm(cfg, graph)
-
-
-def hellings_algorithm_from_pydot(
-    cfg: Union[str, CFG], dot_file: str
-) -> Set[Tuple[str, str, str]]:
-    return hellings_algorithm(cfg, nx_pydot.from_pydot(dot_file))
-
-
-def reachability_with_nonterminal(
-    grammar: Union[str, CFG],
-    graph: MultiDiGraph,
-    start_vertices: Set,
-    end_vertices: Set,
-    target_nonterminal: Variable,
-) -> Set[Tuple[int, int]]:
-    """
-    Computes the reachability information for the specified start and end vertices
-    and the given nonterminal in the context-free grammar.
-
-    Args:
-        grammar(CFG): The context-free grammar (CFG) to use for reachability analysis.
-        graph(MultiDiGraph): The directed graph on which to perform reachability analysis.
-        start_vertices(Set[int]): A set of start vertices for which to compute reachability.
-        end_vertices(Set[int]): A set of end vertices for which to compute reachability.
-        target_nonterminal(Variable): The nonterminal to consider for reachability analysis.
-
-    Returns:
-        Set[Tuple[int, int]]: A set of pairs (start_vertex, end_vertex) representing the reachability
-        information for the specified start and end vertices and the given nonterminal in the CFG.
-    """
-    if isinstance(grammar, str):
-        cfg = cfg_from_text(grammar)
-    reachability = hellings_algorithm(grammar, graph)
-
-    filtered_reachability = {
-        (src, dest)
-        for src, nonterm, dest in reachability
-        if nonterm == target_nonterminal
-        and src in start_vertices
-        and dest in end_vertices
-    }
-
-    return filtered_reachability
