+++ tests/test_cfqp.py
@@ -2,24 +2,26 @@ import pytest
 from pyformlang.cfg import CFG, Variable
 from networkx import MultiDiGraph
 
-from project.graph_utils import hellings_algorithm, reachability_with_nonterminal
+from project.cfqp import hellings, matrix, reachability_with_nonterminal
 from project.graph_utils import create_labeled_two_cycles_graph
 
 
-def test_hellings_algorithm():
+@pytest.mark.parametrize("algo", [hellings, matrix])
+def test_algorithms(algo):
     cfg_text = "S -> a S b | eps"
     cfg = CFG.from_text(cfg_text)
     graph = MultiDiGraph()
     graph.add_edges_from([(0, 1, {"label": "a"}), (1, 2, {"label": "b"})])
 
-    result = hellings_algorithm(cfg, graph)
+    result = algo(cfg, graph)
 
     expected_result = {(1, Variable("b#CNF#"), 2), (0, Variable("a#CNF#"), 1)}
 
     assert result == expected_result
 
 
-def test_reachability_with_nonterminal():
+@pytest.mark.parametrize("algo", ["hellings", "matrix"])
+def test_reachability_with_nonterminal(algo: str):
     cfg_text = """
         S -> A B | B A
         A -> a A b | a b
@@ -44,14 +46,15 @@ def test_reachability_with_nonterminal():
     target_nonterminal = Variable("S")
 
     result = reachability_with_nonterminal(
-        cfg, graph, start_vertices, end_vertices, target_nonterminal
+        cfg, graph, start_vertices, end_vertices, target_nonterminal, algo=algo
     )
 
     expected_result = {(0, 8), (0, 4)}
     assert expected_result == result
 
 
-def test_reachability_with_nonterminal2():
+@pytest.mark.parametrize("algo", ["hellings", "matrix"])
+def test_reachability_with_nonterminal2(algo: str):
     cfg_text = """
         S -> A B
         S -> A S1
@@ -67,7 +70,7 @@ def test_reachability_with_nonterminal2():
     target_nonterminal = Variable("S")
 
     result = reachability_with_nonterminal(
-        cfg, graph, start_vertices, end_vertices, target_nonterminal
+        cfg, graph, start_vertices, end_vertices, target_nonterminal, algo=algo
     )
 
     expected_result = {(0, 0), (0, 3), (1, 0), (1, 3), (2, 0), (2, 3)}
